pyproject.toml
[tool.poetry]
name = "gavel"
version = "0.1.0"
description = "The Gavel - Company Operating System"
authors = ["The Gavel Team <dev@example.com>"] # You can update this later
readme = "README.md"
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.36.0"
supabase = "^2.5.0"
python-dotenv = "^1.0.1"
openai = "^1.34.0"
streamlit-oauth = "^0.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 